simple guess number game
That is a simple game where the user will need to guess a random number, using a random library. 
The game uses a while loop to be able to continue the game giving specific information for the user.
The user is also capable of quit the game when he wishes.